feat: move searcher into internal pkg
