+++ internal/pkg/searcher/noop/client.go
