+++ internal/pkg/searcher/typesense/client.go
