+++ worker/task/process_event_creation_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/frain-dev/convoy"
 	"github.com/frain-dev/convoy/cache"
 	"github.com/frain-dev/convoy/datastore"
+	"github.com/frain-dev/convoy/internal/pkg/searcher"
 	"github.com/frain-dev/convoy/mocks"
 	"github.com/frain-dev/convoy/queue"
-	"github.com/frain-dev/convoy/searcher"
 	"github.com/golang/mock/gomock"
 	"github.com/google/uuid"
 	"github.com/hibiken/asynq"
